Stepping the version number back, and a largly updated Readme/roadmap
Back in the good old days, I thought weave had a feature set worthy of the "40%" mark given by a 0.4.x version tag. With time I've now realized that a version number that high just isn't fitting. So I'm stepping it back for now.

Also updated our goals and milestones in the comments of weave.tl.